app/requirements.txt
Django==3.1.1
django-axes==5.8.0
django-ipware==3.0.1
psycopg2==2.8.6
requests==2.24.0
django-session-timeout==0.1.0
gunicorn==20.0.4